3 Ekim Workshop

+++ src/main/java/com/turkcell/spring/first/repositories/OrderRepository.java
@@ -3,6 +3,6 @@ package com.turkcell.spring.first.repositories;
 import com.turkcell.spring.first.entities.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Short> {
     Order findByOrderId(int orderId);
 }
